Relocate helper test classes (beta)
Same as #1592, change tests to use the helper/utility test classes from
a "neutral" package.
