+++ test/org/zaproxy/zap/extension/ascanrulesBeta/SourceCodeDisclosureWEBINFUnitTest.java
@@ -32,6 +32,7 @@ import org.apache.commons.httpclient.URIException;
 import org.junit.Test;
 import org.parosproxy.paros.core.scanner.Alert;
 import org.parosproxy.paros.network.HttpMessage;
+import org.zaproxy.zap.testutils.NanoServerHandler;
 
 import fi.iki.elonen.NanoHTTPD.IHTTPSession;
 import fi.iki.elonen.NanoHTTPD.Response;
