+++ test/org/zaproxy/zap/testutils/HTTPDTestServer.java
@@ -17,7 +17,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.zaproxy.zap.extension.ascanrulesBeta;
+package org.zaproxy.zap.testutils;
 
 import java.util.ArrayList;
 import java.util.List;
